+++ python/sglang/srt/mem_cache/base_prefix_cache.py
@@ -31,7 +31,7 @@ if TYPE_CHECKING:
         CacheAction,
         ComponentAction,
     )
-    from sglang.srt.mem_cache.unified_cache_components.tree_component import (
+    from sglang.srt.mem_cache.unified_cache.components.tree_component import (
         ComponentType,
     )
