+++ python/sglang/srt/mem_cache/hybrid_cache/hybrid_pool_assembler.py
@@ -27,7 +27,7 @@ from sglang.srt.mem_cache.pool_host.mha import (
     get_mha_host_pool_cls,
 )
 from sglang.srt.mem_cache.pool_host.mla import MLATokenToKVPoolHost
-from sglang.srt.mem_cache.unified_cache_components import ComponentType
+from sglang.srt.mem_cache.unified_cache.components import ComponentType
 
 if TYPE_CHECKING:
     import torch
