+++ python/sglang/srt/mem_cache/unified_cache/tree_core_registry.py
@@ -16,7 +16,7 @@ if TYPE_CHECKING:
     from sglang.srt.mem_cache.unified_cache.unified_tree_core_interface import (
         UnifiedTreeCoreInterface,
     )
-    from sglang.srt.mem_cache.unified_cache_components import TreeComponent
+    from sglang.srt.mem_cache.unified_cache.components import TreeComponent
 
 TreeCoreFactory = Callable[
     ["CacheInitParams", "dict[ComponentType, TreeComponent]"],
